Fix inability to write outside of the presentation
Since universe is transformed (so taken as a clipping context) we cannot just
use `overflow: display`. So we add a padding of 2000px around the presentation,
which should be enough.

The 2000 value is duplicated between css and code. We add comments to link them
together. It would be better to have it defined in a single place, but...